// libs/affiliate-billing/preferences/filds-view/src/lib/enums/billing-preferences-field.enum.ts
export enum BillingPreferencesFieldEnum {
    Frequency = 'frequency',
    GenerationInvoice = 'generationInvoice',
    PaymentTerms = 'paymentTerms',
    BeneficiaryName = 'beneficiaryName',
    BeneficiaryAddress = 'beneficiaryAddress',
    BillingEmail = 'billingEmail',
    TaxId = 'taxId',
    Vat = 'vat'
}